C2-36-39 Auther added, user foreign model, custom template tag added, blcok post who isn't staff user

+++ src/posts/views.py
@@ -1,6 +1,6 @@
 from urllib.parse import quote_plus
 from django.contrib import messages
-from django.http import HttpResponse, HttpResponseRedirect
+from django.http import HttpResponse, HttpResponseRedirect, Http404
 from django.shortcuts import render, get_object_or_404, redirect
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
 
@@ -9,9 +9,13 @@ from .models import Post
 from .forms import PostForm
 
 def post_create(request):
+    if request.user.is_staff or not request.user.is_superuser:
+        raise Http404
+
     form = PostForm(request.POST or None, request.FILES or None)
     if form.is_valid():
         print(form.cleaned_data.get("title"))
+        instance.user = request.user
         instance = form.save(commit=False)
         instance.save()
         # message success
@@ -58,6 +62,8 @@ def post_list(request):# list item
 
 
 def post_update(request, slug=None):
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404 
     instance = get_object_or_404(Post, slug=slug)
     form = PostForm(request.POST or None, request.FILES or None, instance=instance)
     # form is posted
@@ -75,6 +81,8 @@ def post_update(request, slug=None):
     return render(request, "post_form.html", context)
 
 def post_delete(request, slug=None):
+    if request.user.is_staff or not request.user.is_superuser:
+        raise Http404    
     instance = get_object_or_404(Post, slug=slug)
     instance.delete()
     messages.success(request, "Successfully Delete")
